flowgrind: daily commit - various minor and major bugfixes, unified DEBUG_MSG
git-svn-id: svn+ssh://svn.umic-mesh.net/umic-mesh/projects/flowgrind/branches/flowgrind-next@6359 3d4b872d-e00e-0410-a2ce-a5c2cda4747c

+++ src/common.h
@@ -130,11 +130,11 @@ struct _report
 
 	long long bytes_read;
 	long long bytes_written;
-
-	long request_blocks_read;
-	long request_blocks_written;
-	long response_blocks_read;
-	long response_blocks_written;
+	/* TODO: figure out if long is needed here */
+	int request_blocks_read;
+	int request_blocks_written;
+	int response_blocks_read;
+	int response_blocks_written;
 
 	double rtt_min, rtt_max, rtt_sum;
 	double iat_min, iat_max, iat_sum;
